prepare functions for parallel use

+++ Matrix.cpp
@@ -1,28 +1,28 @@
 #include "Matrix.h"
 
-Matrix createZeroMatrix(int rows, int columns) {
+Matrix createZeroMatrix(size_t rows, size_t columns) {
 	Matrix matrix;
 	matrix.rowsCount = rows;
 	matrix.columnsCount = columns;
 	matrix.matrix = new int*[matrix.rowsCount];
-	for (int i = 0; i < matrix.rowsCount; ++i) {
+	for (size_t i = 0; i < matrix.rowsCount; ++i) {
 		matrix.matrix[i] = new int[matrix.columnsCount];
-		for (int j = 0; j < matrix.columnsCount; ++j) {
+		for (size_t j = 0; j < matrix.columnsCount; ++j) {
 			matrix.matrix[i][j] = 0;
 		}
 	}
 	return matrix;
 }
 
-Matrix createRandomMatrix(int rows, int columns) {
+Matrix createRandomMatrix(size_t rows, size_t columns) {
 	Matrix matrix;
 	matrix.rowsCount = rows;
 	matrix.columnsCount = columns;
 
 	matrix.matrix = new int*[matrix.rowsCount];
-	for (int i = 0; i < matrix.rowsCount; ++i) {
+	for (size_t i = 0; i < matrix.rowsCount; ++i) {
 		matrix.matrix[i] = new int[matrix.columnsCount];
-		for (int j = 0; j < matrix.columnsCount; ++j) {
+		for (size_t j = 0; j < matrix.columnsCount; ++j) {
 			matrix.matrix[i][j] = rand();
 		}
 	}
@@ -30,8 +30,8 @@ Matrix createRandomMatrix(int rows, int columns) {
 }
 void printMatrix(Matrix matrix) {
 	cout << "\n";
-	for (int i = 0; i < matrix.rowsCount; ++i) {
-		for (int j = 0; j < matrix.columnsCount; ++j) {
+	for (size_t i = 0; i < matrix.rowsCount; ++i) {
+		for (size_t j = 0; j < matrix.columnsCount; ++j) {
 			cout << matrix.matrix[i][j] << '\t';
 		}
 		cout << "\n";
